MAT-6790 Created a new controller for VSAC Fhir Terminology related resources
