+++ src/test/java/gov/cms/madie/terminology/controller/VsacControllerMvcTest.java
@@ -2,12 +2,10 @@ package gov.cms.madie.terminology.controller;
 
 import ca.uhn.fhir.context.FhirContext;
 import generated.vsac.nlm.nih.gov.RetrieveMultipleValueSetsResponse;
-import gov.cms.madie.models.measure.ManifestExpansion;
 import gov.cms.madie.terminology.dto.QdmValueSet;
 import gov.cms.madie.terminology.dto.ValueSetsSearchCriteria;
 import gov.cms.madie.terminology.helpers.TestHelpers;
 import gov.cms.madie.terminology.models.UmlsUser;
-import gov.cms.madie.terminology.service.FhirTerminologyService;
 import gov.cms.madie.terminology.service.VsacService;
 import jakarta.xml.bind.JAXBContext;
 import jakarta.xml.bind.JAXBException;
@@ -27,7 +25,6 @@ import org.springframework.web.reactive.function.client.WebClientResponseExcepti
 import java.io.File;
 import java.io.IOException;
 import java.security.Principal;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -48,12 +45,7 @@ public class VsacControllerMvcTest {
 
   @MockBean private VsacService vsacService;
 
-  @MockBean private FhirTerminologyService fhirTerminologyService;
-
   @Autowired private MockMvc mockMvc;
-
-  private UmlsUser umlsUser;
-  private final List<ManifestExpansion> mockManifests = new ArrayList<>();
 
   private RetrieveMultipleValueSetsResponse svsValueSet;
   private ValueSet fhirValueSet;
@@ -62,18 +54,6 @@ public class VsacControllerMvcTest {
 
   @BeforeEach
   public void setup() throws JAXBException, IOException {
-    umlsUser = UmlsUser.builder().apiKey(TEST_API_KEY).harpId(TEST_USER).build();
-    mockManifests.add(
-        ManifestExpansion.builder()
-            .fullUrl("https://cts.nlm.nih.gov/fhir/Library/ecqm-update-4q2017-eh")
-            .id("ecqm-update-4q2017-eh")
-            .build());
-    mockManifests.add(
-        ManifestExpansion.builder()
-            .fullUrl("https://cts.nlm.nih.gov/fhir/Library/mu2-update-2012-10-25")
-            .id("mu2-update-2012-10-25")
-            .build());
-
     File file = TestHelpers.getTestResourceFile("/value-sets/svs_office_visit.xml");
     JAXBContext jaxbContext = JAXBContext.newInstance(RetrieveMultipleValueSetsResponse.class);
     Unmarshaller jaxbUnmarshaller = jaxbContext.createUnmarshaller();
@@ -247,47 +227,6 @@ public class VsacControllerMvcTest {
                     .with(user(TEST_USR))
                     .with(csrf())
                     .content(searchCriteria)
-                    .contentType(MediaType.APPLICATION_JSON_VALUE))
-            .andExpect(status().isUnauthorized())
-            .andReturn();
-    assertThat(result.getResponse().getStatus(), is(equalTo(401)));
-  }
-
-  @Test
-  void testGetManifestsSuccessfullyMvc() throws Exception {
-    Principal principal = mock(Principal.class);
-    when(principal.getName()).thenReturn(TEST_USER);
-    when(vsacService.findByHarpId(anyString())).thenReturn(Optional.ofNullable(umlsUser));
-    when(fhirTerminologyService.getManifests(any(UmlsUser.class))).thenReturn(mockManifests);
-    MvcResult result =
-        mockMvc
-            .perform(
-                MockMvcRequestBuilders.get("/vsac/manifest-list")
-                    .with(user(TEST_USR))
-                    .with(csrf())
-                    .contentType(MediaType.APPLICATION_JSON_VALUE))
-            .andExpect(status().isOk())
-            .andReturn();
-    assertThat(result.getResponse().getStatus(), is(equalTo(200)));
-    String content = result.getResponse().getContentAsString();
-    verify(fhirTerminologyService, times(1)).getManifests(any(UmlsUser.class));
-    assertThat(
-        content,
-        containsString(
-            "[{\"fullUrl\":\"https://cts.nlm.nih.gov/fhir/Library/ecqm-update-4q2017-eh\",\"id\":\"ecqm-update-4q2017-eh\"},{\"fullUrl\":\"https://cts.nlm.nih.gov/fhir/Library/mu2-update-2012-10-25\",\"id\":\"mu2-update-2012-10-25\"}]"));
-  }
-
-  @Test
-  void testUnAuthorizedUmlsUserWhileFetchingManifestsMvc() throws Exception {
-    Principal principal = mock(Principal.class);
-    when(principal.getName()).thenReturn(TEST_USER);
-    when(vsacService.findByHarpId(anyString())).thenReturn(Optional.empty());
-    MvcResult result =
-        mockMvc
-            .perform(
-                MockMvcRequestBuilders.get("/vsac/manifest-list")
-                    .with(user(TEST_USR))
-                    .with(csrf())
                     .contentType(MediaType.APPLICATION_JSON_VALUE))
             .andExpect(status().isUnauthorized())
             .andReturn();
