+++ src/test/java/gov/cms/madie/terminology/controller/VsacControllerTest.java
@@ -1,9 +1,5 @@
 package gov.cms.madie.terminology.controller;
 
-import gov.cms.madie.models.measure.ManifestExpansion;
-import gov.cms.madie.terminology.exceptions.VsacUnauthorizedException;
-import gov.cms.madie.terminology.service.FhirTerminologyService;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -19,7 +15,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.security.Principal;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -34,36 +29,14 @@ import gov.cms.madie.terminology.service.VsacService;
 
 @ExtendWith(MockitoExtension.class)
 public class VsacControllerTest {
+  @Mock
+  private VsacService vsacService;
 
-  @Mock private VsacService vsacService;
-  @Mock private FhirTerminologyService fhirTerminologyService;
   @InjectMocks private VsacController vsacController;
 
   private static final String TEST = "test";
   private static final String TEST_USER = "test.user";
-
-  private UmlsUser umlsUser;
-
-  private static final String TEST_HARP_ID = "te$tHarpId";
-  private static final String TEST_API_KEY = "te$tKey";
   private static final String FHIR_DATA_MODEL = "FHIR";
-
-  private final List<ManifestExpansion> mockManifests = new ArrayList<>();
-
-  @BeforeEach
-  public void setUp() {
-    umlsUser = UmlsUser.builder().apiKey(TEST_API_KEY).harpId(TEST_HARP_ID).build();
-    mockManifests.add(
-        ManifestExpansion.builder()
-            .fullUrl("https://cts.nlm.nih.gov/fhir/Library/ecqm-update-4q2017-eh")
-            .id("ecqm-update-4q2017-eh")
-            .build());
-    mockManifests.add(
-        ManifestExpansion.builder()
-            .fullUrl("https://cts.nlm.nih.gov/fhir/Library/mu2-update-2012-10-25")
-            .id("mu2-update-2012-10-25")
-            .build());
-  }
 
   @Test
   void testGetValueSetFailWhenGettingValueSetFailed() {
@@ -160,25 +133,5 @@ public class VsacControllerTest {
     ResponseEntity<Boolean> response = vsacController.checkUserLogin(principal);
 
     assertEquals(response.getStatusCode(), HttpStatus.UNAUTHORIZED);
-  }
-
-  @Test
-  void testGetManifestsSuccessfully() {
-    Principal principal = mock(Principal.class);
-    when(principal.getName()).thenReturn(TEST_USER);
-    when(vsacService.findByHarpId(anyString())).thenReturn(Optional.ofNullable(umlsUser));
-    when(fhirTerminologyService.getManifests(any(UmlsUser.class))).thenReturn(mockManifests);
-    ResponseEntity<List<ManifestExpansion>> response = vsacController.getManifests(principal);
-    assertEquals(response.getStatusCode(), HttpStatus.OK);
-    assertEquals(response.getBody(), mockManifests);
-  }
-
-  @Test
-  void testUnAuthorizedUmlsUserWhileFetchingManifests() {
-    Principal principal = mock(Principal.class);
-    when(principal.getName()).thenReturn(TEST_USER);
-    when(vsacService.findByHarpId(anyString()))
-        .thenReturn(Optional.ofNullable(UmlsUser.builder().build()));
-    assertThrows(VsacUnauthorizedException.class, () -> vsacController.getManifests(principal));
   }
 }
